requirements.txt
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
humanize==4.6.0
idna==3.4
joblib==1.2.0
mysql-connector==2.2.9
numpy==1.24.2
prettytable==3.6.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
soupsieve==2.4
threadpoolctl==3.1.0
urllib3==1.26.15
wcwidth==0.2.6
